Extracted twitchclient for retrieveUser
Extrated the retrieve operation for User on the twitch client
Fixed the compile error
Moved the implementation of type classes into its companion objects
Still on the fence on using the ZIO runtime on around the code instead
of abstracting it on an `Handler`
