+++ src/main/scala/hazzlenut/api/Authentication.scala
@@ -1,24 +1,23 @@
 package hazzlenut.api
 
 import akka.actor.ActorRef
-import akka.pattern.ask
-import akka.http.scaladsl.Http
-import akka.http.scaladsl.model.headers.OAuth2BearerToken
-import akka.http.scaladsl.model.{HttpRequest, StatusCodes, Uri}
+import akka.http.scaladsl.model.{StatusCodes, Uri}
 import akka.http.scaladsl.server.Directives._
-import akka.http.scaladsl.unmarshalling.Unmarshal
+import akka.pattern.ask
 import akka.util.Timeout
+import hazzlenut.errors.HazzlenutError
 import hazzlenut.handler.AuthenticationHandler
 import hazzlenut.services.twitch.TokenGuardian.Authenticated
 import hazzlenut.services.twitch.TokenHolder.{AskAccessToken, ReplyAccessToken}
-import hazzlenut.services.twitch.model.{TwitchReply, User}
+import hazzlenut.services.twitch.TwitchClient.dsl.retrieveUser
+import hazzlenut.util.ZIORuntime.runtime
+import scalaz.zio.ZIO
 
-import concurrent.duration._
+import scala.concurrent.duration._
 
 object Authentication {
 
   def publicRoute(implicit tokenGuardian: ActorRef) = {
-    import hazzlenut.services.twitch.TwitchZIO._
     route
   }
 
@@ -82,25 +81,17 @@ object Authentication {
                   implicit val m = materializer
                   implicit val timeout = Timeout(5 seconds)
 
-                  onSuccess(
-                    (tokenGuardian ? AskAccessToken).mapTo[ReplyAccessToken]
-                  ) { token =>
-                    onSuccess(
-                      Http().singleRequest(
-                        HttpRequest(uri = "https://api.twitch.tv/helix/users")
-                          .addCredentials(
-                            OAuth2BearerToken(token.accessToken.accessToken)
-                          )
+                  onSuccess(for {
+                    token <- (tokenGuardian ? AskAccessToken)
+                      .mapTo[ReplyAccessToken]
+                    user <- runtime.unsafeRunToFuture(
+                      retrieveUser[ZIO[Any, HazzlenutError, ?]](
+                        token.accessToken
                       )
-                    ) { user =>
-                      onSuccess(Unmarshal(user.entity).to[TwitchReply[User]]) {
-                        userInfo =>
-                          complete(s"user ${userInfo.data.head}")
-                      }
-
-                    }
+                    )
+                  } yield user) { u =>
+                    complete(s"user ${u}")
                   }
-
                 }
             }
           }
