+++ src/main/scala/hazzlenut/handler/AuthenticationHandler.scala
@@ -3,7 +3,9 @@ package hazzlenut.handler
 import akka.actor.ActorSystem
 import akka.stream.Materializer
 import hazzlenut.errors.HazzlenutError
-import hazzlenut.services.twitch.AccessToken
+import hazzlenut.services.twitch.{AccessToken, Authenticate}
+import scalaz.zio.ZIO
+import scalaz.zio.interop.catz._
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -16,11 +18,13 @@ trait AuthenticationHandler {
                                 ec: ExecutionContext,
                                 mat: Materializer): Future[AccessToken]
 
-  def refreshToken(code: String)(implicit system: ActorSystem,
-                                ec: ExecutionContext,
-                                mat: Materializer): Future[Either[HazzlenutError, AccessToken]]
+  def refreshToken(code: String)(
+    implicit system: ActorSystem,
+    ec: ExecutionContext,
+    mat: Materializer
+  ): Future[Either[HazzlenutError, AccessToken]]
 
-  def reAuthenticate(): Either[HazzlenutError,Unit]
+  def reAuthenticate(): Either[HazzlenutError, Unit]
 }
 
 object AuthenticationHandler {
@@ -49,7 +53,51 @@ object AuthenticationHandler {
     ): Future[Either[HazzlenutError, AccessToken]] =
       authentication.refreshToken(refreshToken)
 
-    def reAuthenticate(implicit authenticationHandler: AuthenticationHandler): Either[HazzlenutError, Unit] =
+    def reAuthenticate(
+      implicit authenticationHandler: AuthenticationHandler
+    ): Either[HazzlenutError, Unit] =
       authenticationHandler.reAuthenticate()
+  }
+
+  implicit val authHandlerZIO = new AuthenticationHandler {
+    import hazzlenut.util.ZIORuntime._
+
+    override def getAuthUrl(implicit system: ActorSystem,
+                            ec: ExecutionContext,
+                            mat: Materializer): Future[Option[String]] = {
+      val getUrl =
+        Authenticate.getUrlToAuthenticate[ZIO[Any, HazzlenutError, ?]]
+
+      runtime.unsafeRunToFuture(getUrl)
+    }
+
+    override def obtainOAuth(code: String)(
+      implicit system: ActorSystem,
+      ec: ExecutionContext,
+      mat: Materializer
+    ): Future[AccessToken] = {
+      val getOAuthToken =
+        Authenticate.authenticate[ZIO[Any, HazzlenutError, ?]]
+
+      runtime.unsafeRunToFuture(getOAuthToken.run(code))
+    }
+
+    override def refreshToken(code: String)(
+      implicit system: ActorSystem,
+      ec: ExecutionContext,
+      mat: Materializer
+    ): Future[Either[HazzlenutError, AccessToken]] = {
+      val refreshToken =
+        Authenticate.refresh[ZIO[Any, HazzlenutError, ?]]
+
+      runtime.unsafeRunToFuture(refreshToken.run(code).either)
+    }
+
+    override def reAuthenticate(): Either[HazzlenutError, Unit] = {
+      val reauthentication =
+        Authenticate.reAuthenticate[ZIO[Any, HazzlenutError, ?]]
+
+      runtime.unsafeRun(reauthentication)
+    }
   }
 }
