+++ src/main/scala/hazzlenut/services/twitch/Configuration.scala
@@ -1,15 +1,16 @@
 package hazzlenut.services.twitch
 
+import cats.Monad
 import cats.implicits._
-import cats.{Monad, MonadError}
 import hazzlenut.errors.HazzlenutError
 import hazzlenut.errors.HazzlenutError.{InvalidConfiguration, MonadErrorHazzlenut}
 import hazzlenut.services.twitch.Configuration.Config
+import hazzlenut.util.MapGetterValidation.{ConfigurationValidation, _}
+import scalaz.zio.ZIO
 
 trait Configuration[F[_]] {
   import hazzlenut.util.MapGetterValidation._
-  def get(implicit m: Monad[F],
-          mError: MonadErrorHazzlenut[F]): F[Config] =
+  def get(implicit m: Monad[F], mError: MonadErrorHazzlenut[F]): F[Config] =
     for {
       configOrValidationError <- pureAsync {
         (
@@ -43,6 +44,18 @@ trait Configuration[F[_]] {
 }
 
 object Configuration {
+
+  def apply[F[_]](implicit configuration: Configuration[F]): Configuration[F] =
+    configuration
+
+  case class Config(clientId: String,
+                    clientSecret: String,
+                    redirectUri: String,
+                    tokenUrl: String,
+                    authorizeUrl: String,
+                    siteUrl: String,
+                    scopes: String*)
+
   object Config {
     def apply(clientId: String,
               clientSecret: String,
@@ -62,21 +75,37 @@ object Configuration {
       )
   }
 
-  case class Config(clientId: String,
-                    clientSecret: String,
-                    redirectUri: String,
-                    tokenUrl: String,
-                    authorizeUrl: String,
-                    siteUrl: String,
-                    scopes: String*)
-
-  def apply[F[_]](implicit configuration: Configuration[F]): Configuration[F] =
-    configuration
-
   object dsl {
     def get[F[_]: Monad: MonadErrorHazzlenut](
       implicit configuration: Configuration[F]
     ): F[Configuration.Config] =
       configuration.get
   }
+
+  implicit val configurationZIO =
+    new Configuration[ZIO[Any, HazzlenutError, ?]] {
+
+      override def pureAsync(
+        f: => ConfigurationValidation[Configuration.Config]
+      ): ZIO[Any, HazzlenutError, ConfigurationValidation[
+        Configuration.Config
+      ]] = ZIO.succeedLazy(f)
+
+      override def getConfig(): (ConfigurationValidation[String],
+                                 ConfigurationValidation[String],
+                                 ConfigurationValidation[String],
+                                 ConfigurationValidation[String],
+                                 ConfigurationValidation[String],
+                                 ConfigurationValidation[String],
+                                 ConfigurationValidation[String]) =
+        (
+          sys.env.getMandatory("clientId"),
+          sys.env.getMandatory("clientSecret"),
+          sys.env.getMandatory("redirectUri"),
+          sys.env.getMandatory("tokenUrl"),
+          sys.env.getMandatory("authorizeUrl"),
+          sys.env.getMandatory("siteUrl"),
+          sys.env.getMandatory("scopes")
+        )
+    }
 }
