+++ src/main/scala/hazzlenut/services/twitch/Reauthentication.scala
@@ -1,6 +1,10 @@
 package hazzlenut.services.twitch
 
 import hazzlenut.errors.HazzlenutError
+import hazzlenut.errors.HazzlenutError.ThrowableError
+import scalaz.zio.ZIO
+
+import scala.util.Try
 
 trait Reauthentication[F[_]] {
   def requestAuthenticationFromUser(): F[Either[HazzlenutError,Unit]]
@@ -12,5 +16,25 @@ object Reauthentication{
   object dsl {
     def reauthenticateWithUser[F[_]: Reauthentication](): F[Either[HazzlenutError, Unit]] =
       Reauthentication[F].requestAuthenticationFromUser()
+  }
+
+  implicit val ReauthenticateZIO: Reauthentication[ZIO[Any, HazzlenutError, ?]] = new Reauthentication[ZIO[Any, HazzlenutError, ?]] {
+    override def requestAuthenticationFromUser()
+    : ZIO[Any, HazzlenutError, Either[HazzlenutError, Unit]] = {
+      import java.awt.Desktop
+      import java.net.URI
+
+      ZIO.fromTry {
+        Try {
+          if (Desktop.isDesktopSupported() && Desktop
+            .getDesktop()
+            .isSupported(Desktop.Action.BROWSE)) {
+            Desktop
+              .getDesktop()
+              .browse(new URI("http://localhost:8000/oauth/login"))
+          }
+        }
+      }.mapError{throwable => ThrowableError(throwable)}.either
+    }
   }
 }
